refactor: :recycle: rename useModifiedTransferAccounts hook, improve types for landing company
